requirements.txt
flask
PyPDF2
python-docx
nltk
scikit-learn
